Refactoring index migration task
