+++ modules/joex/src/main/scala/docspell/joex/fts/MigrateCtx.scala
@@ -1,0 +1,13 @@
+package docspell.joex.fts
+
+import docspell.common.Logger
+import docspell.joex.Config
+import docspell.store.Store
+import docspell.ftsclient.FtsClient
+
+case class MigrateCtx[F[_]](
+    cfg: Config.FullTextSearch,
+    store: Store[F],
+    fts: FtsClient[F],
+    logger: Logger[F]
+)
