+++ modules/joex/src/main/scala/docspell/joex/fts/package.scala
@@ -1,0 +1,9 @@
+package docspell.joex
+
+import cats.data.Kleisli
+
+package object fts {
+
+  type MigrationTask[F[_]] = Kleisli[F, MigrateCtx[F], Unit]
+
+}
